+++ src/Application/VsSettings.cpp
@@ -78,3 +78,15 @@ QMap<QString, QVariant> VsSettings::speakerSettings()
 
 	return ttsSettings;
 }
+
+QMap<QString, QVariant> VsSettings::quizSettings()
+{
+	QMap<QString, QVariant> quizSettings;
+
+	quizSettings["displayQuizAnswerStatus"]	= value( "displayQuizAnswerStatus", "Quiz", QVariant( false ) );
+	quizSettings["randomiizeWords"]			= value( "randomiizeWords", "Quiz", QVariant( true ) );
+	quizSettings["displayTimer"]			= value( "displayTimer", "Quiz", QVariant( false ) );
+	quizSettings["timerDefaultTime"]		= value( "volume", "Quiz", QVariant( 1800 ) );
+
+	return quizSettings;
+}