+++ src/Widget/Quiz/QuizParametersWidget.cpp
@@ -14,15 +14,21 @@
 #include "GlobalTypes.h"
 #include "Entity/VocabularyMetaInfo.h"
 #include "Entity/VocabularyGroup.h"
+#include "Application/VsSettings.h"
 
 QuizParametersWidget::QuizParametersWidget( QWidget *parent ) :
 	QWidget( parent ),
     ui( new Ui::QuizParametersWidget )
 {
     ui->setupUi( this );
+    quizSettings		= VsSettings::instance()->quizSettings();
+    bool displayTimer	= quizSettings["displayQuizAnswerStatus"].toBool();
+
+    ui->chkRandomize->setChecked( quizSettings["randomiizeWords"].toBool() );
+    ui->chkTimer->setChecked( displayTimer );
 
     initGroups();
-    initTimer( false );
+    initTimer( displayTimer );
     setDirection();
 
     connect(
@@ -58,8 +64,8 @@ void QuizParametersWidget::setMetaInfo( VocabularyMetaInfoPtr metaInfo )
 
 void QuizParametersWidget::initTimer( bool on )
 {
-	ui->leTimer->setValidator( new QIntValidator( 0, 3600, this ) );
-	ui->leTimer->setText( QString::number( 1800 ) );
+	QTime timer( 0, 0 );
+	ui->teTimer->setTime( timer.addSecs( quizSettings["timerDefaultTime"].toInt() ) );
 
 	if ( on ) {
 		ui->frmTime->show();
@@ -123,7 +129,8 @@ int QuizParametersWidget::time()
 {
 	int time	= 0;
 	if ( ui->chkTimer->isChecked() ) {
-		time	= ui->leTimer->text().toInt();
+		QTime tdt	= ui->teTimer->time();
+		time		= ( tdt.hour() * 3600 ) + ( tdt.minute() * 60 ) + tdt.second();
 	}
 
 	return time;